Added Platform project in Game. Currently compiling and the EventManager seems to be working but only the Jet and Missiles are INSTANTIATED.

+++ Source/Engine/Renderer/Renderer.cpp
@@ -101,8 +101,8 @@ namespace kiko
 		vec2 size = texture->GetSize();
 
 		SDL_Rect dest;
-		dest.x = (x - (size.x * 0.5f));
-		dest.y = (y - (size.y * 0.5f));
+		dest.x = (int)(x - (size.x * 0.5f));
+		dest.y = (int)(y - (size.y * 0.5f));
 
 		dest.w = (int)size.x;
 		dest.h = (int)size.y;
